requirements.txt
# Compatible versions for voice AI agent
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
pyannote.audio==3.1.1
pyannote.core<6.0
pyannote.metrics<4.0
pyannote.pipeline<4.0
pandas>=1.3.0
python-dotenv>=0.19.0
soundfile>=0.10.0
numpy<2.0
huggingface-hub>=0.13.0
